[UPDATE] - move the router enum to be located with the app drawer (i.e. the screen that is dependent on it) and move setup functions for shared preferences and the notification system from the common file to the new setup file for holding file to be used for setting up the app
